Added migration fields

+++ database/migrations/2020_03_25_125122_create_alert_translations_table.php
@@ -17,7 +17,7 @@ class CreateAlertTranslationsTable extends Migration
             $table->id();
             $table->string('title');
             $table->string('locale');
-            $table->integer('alert_id')->unsigned()->index();
+            $table->bigInteger('alert_id')->unsigned()->index();
             $table->text('content')->nullable();
             $table->timestamps();
